getByte now returns int. Lots of debug. One bug fix.
Some external code was having issues converting the byte returned by
getByte to an int properly. I'm not sure this is the right answer but
we'll try it for now.

Trying to figure out what is going on with some subchunk oddities from
a different project.

ByteChunkSPI was always writing to array offset 0 on copyTo.

+++ src/main/java/net/darkmist/chunks/AbstractChunkSPI.java
@@ -70,7 +70,7 @@ public abstract class AbstractChunkSPI implements ChunkSPI
 	 * @throws IndexOutOfBoundsException if <code>off</code> is greater than <code>Integer.MAX_VALUE</code>.
 	 */
 	@Override
-	public abstract byte getByte(long off);
+	public abstract int getByte(long off);
 
 	@Override
 	@SuppressWarnings("PMD.AvoidUsingShortType")
@@ -154,7 +154,7 @@ public abstract class AbstractChunkSPI implements ChunkSPI
 	{
 		int end = Math.addExact(arrayOff, len);
 		for(;chunkOff<size&&arrayOff<end;chunkOff++,arrayOff++)
-			bytes[arrayOff] = getByte(chunkOff);
+			bytes[arrayOff] = (byte)getByte(chunkOff);
 		return bytes;
 	}
 }
