+++ src/main/java/net/darkmist/chunks/ChunkSPI.java
@@ -11,7 +11,7 @@ public interface ChunkSPI
 	 * @return Value at the specified offset.
 	 * @throws IndexOutOfBoundsException if <code>off</code> is negative or greater then or equal to the size.
 	 */
-	public byte getByte(long off);
+	public int getByte(long off);
 
 	@SuppressWarnings("PMD.AvoidUsingShortType")
 	public short getShort(long off, ByteOrder order);
